requirements.txt
pyyaml
docopt
dpath
